Crate mapsengine1

Features

Handle the following Resources with ease from the central hub ...

assets (get, list, list parents and list permissions)

layers (cancelprocessing, create, delete, get, getpublished, list, listpublished, list parents, patch, batchdelete permissions, batchupdate permissions, list permissions, process, publish and unpublish)

maps (create, delete, get, getpublished, list, listpublished, patch, batchdelete permissions, batchupdate permissions, list permissions, publish and unpublish)

projects (create icons, get icons, list icons and list)

raster collections (cancelprocessing, create, delete, get, list, list parents, patch, batchdelete permissions, batchupdate permissions, list permissions, process, batchdelete rasters, batchinsert rasters and list rasters)

rasters (delete, insert files, get, list, list parents, patch, batchdelete permissions, batchupdate permissions, list permissions, process and upload)

tables (create, delete, batchdelete features, batchinsert features, batchpatch features, get features, list features, insert files, get, list, list parents, patch, batchdelete permissions, batchupdate permissions, list permissions, process and upload)

Everything else about the Maps Engine API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.tables().features_batch_insert(...).doit()
let r = hub.tables().list(...).doit()
let r = hub.tables().patch(...).doit()
let r = hub.tables().get(...).doit()
let r = hub.tables().files_insert(...).doit()
let r = hub.tables().features_batch_delete(...).doit()
let r = hub.tables().permissions_batch_delete(...).doit()
let r = hub.tables().permissions_batch_update(...).doit()
let r = hub.tables().features_list(...).doit()
let r = hub.tables().process(...).doit()
let r = hub.tables().upload(...).doit()
let r = hub.tables().delete(...).doit()
let r = hub.tables().parents_list(...).doit()
let r = hub.tables().features_batch_patch(...).doit()
let r = hub.tables().permissions_list(...).doit()
let r = hub.tables().features_get(...).doit()
let r = hub.tables().create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate mapsengine1;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.tables().list()
             .tags("sea")
             .search("nonumy")
             .role("dolores")
             .project_id("gubergren")
             .processing_status("sadipscing")
             .page_token("aliquyam")
             .modified_before("ea")
             .modified_after("no")
             .max_results(80)
             .creator_email("justo")
             .created_before("et")
             .created_after("et")
             .bbox("diam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

AcquisitionTime
Acquisition time represents acquired time of a raster.

This type is not used in any activity, and only used as part of another schema.

Asset
An asset is any Google Maps Engine resource that has a globally unique ID. Assets include maps, layers, vector tables, raster collections, and rasters. Projects and features are not considered assets.

More detailed information about an asset can be obtained by querying the asset's particular endpoint.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

parents.list (none)

list (none)

permissions.list (none)

get (response)

AssetGetMethodBuilder
Return metadata for a particular asset.

AssetListMethodBuilder
Return all assets readable by the current user.

AssetMethodsBuilder
A builder providing access to all methods supported on asset resources. It is not used directly, but through the MapsEngine hub.

AssetParentListMethodBuilder
Return all parent ids of the specified asset.

AssetPermissionListMethodBuilder
Return all of the permissions for the specified asset.

AssetsListResponse
The response returned by a call to resources.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

Border
Border in line style. Both color and width are required.

This type is not used in any activity, and only used as part of another schema.

Color
Basic color used in styling.

This type is not used in any activity, and only used as part of another schema.

Datasource
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Datasources
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

The contained type is Vec<Datasource>.

DisplayRule
A display rule of the vector style.

This type is not used in any activity, and only used as part of another schema.

Feature
A feature within a table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

features.get (response)

FeatureInfo
A feature info contains information about individual feature.

This type is not used in any activity, and only used as part of another schema.

FeaturesBatchDeleteRequest
The request sent to features.BatchDelete.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

features.batchDelete (request)

FeaturesBatchInsertRequest
The request sent to features.Insert.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

features.batchInsert (request)

FeaturesBatchPatchRequest
The request sent to features.BatchPatch.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

features.batchPatch (request)

FeaturesListResponse
The response returned by a call to features.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

features.list (response)

File
A single File, which is a component of an Asset.

This type is not used in any activity, and only used as part of another schema.

Filter
Conditions for filtering features.

This type is not used in any activity, and only used as part of another schema.

GeoJsonGeometry
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeoJsonGeometryCollection
A heterogenous collection of GeoJsonGeometry objects.

This type is not used in any activity, and only used as part of another schema.

GeoJsonLineString
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeoJsonMultiLineString
Multi Line String

This type is not used in any activity, and only used as part of another schema.

GeoJsonMultiPoint
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeoJsonMultiPolygon
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeoJsonPoint
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeoJsonPolygon
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeoJsonPosition
A position represents a geographical position as an array containing a longitude and a latitude, and optionally an altitude, in that order. All Geometry objects make use of positions to represent geometries as nested arrays. The structure of the array is governed by the type of the geometry.

This type is not used in any activity, and only used as part of another schema.

The contained type is Vec<f64>.

GeoJsonProperties
The properties associated with a feature.

This type is not used in any activity, and only used as part of another schema.

Icon
An icon is a user-uploaded image that can be used to style point geometries.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

icons.create (request|response)

icons.get (response)

IconStyle
Style for icon, this is part of point style.

This type is not used in any activity, and only used as part of another schema.

IconsListResponse
The response returned by a call to icons.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

icons.list (response)

LabelStyle
Text label style.

This type is not used in any activity, and only used as part of another schema.

LatLngBox
A rectangular geographic bounds.

This type is not used in any activity, and only used as part of another schema.

The contained type is Vec<f64>.

Layer
A Layer combines multiple datasources, with styling information, for presentation on a map.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete (none)

get (response)

cancelProcessing (none)

publish (none)

permissions.batchDelete (none)

listPublished (none)

create (request|response)

patch (request)

permissions.batchUpdate (none)

permissions.list (none)

list (none)

process (none)

parents.list (none)

getPublished (none)

unpublish (none)

LayerCancelProcessingMethodBuilder
Cancel processing on a layer asset.

LayerCreateMethodBuilder
Create a layer asset.

LayerDeleteMethodBuilder
Delete a layer.

LayerGetMethodBuilder
Return metadata for a particular layer.

LayerGetPublishedMethodBuilder
Return the published metadata for a particular layer.

LayerListMethodBuilder
Return all layers readable by the current user.

LayerListPublishedMethodBuilder
Return all published layers readable by the current user.

LayerMethodsBuilder
A builder providing access to all methods supported on layer resources. It is not used directly, but through the MapsEngine hub.

LayerParentListMethodBuilder
Return all parent ids of the specified layer.

LayerPatchMethodBuilder
Mutate a layer asset.

LayerPermissionBatchDeleteMethodBuilder
Remove permission entries from an already existing asset.

LayerPermissionBatchUpdateMethodBuilder
Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

LayerPermissionListMethodBuilder
Return all of the permissions for the specified asset.

LayerProcesMethodBuilder
Process a layer asset.

LayerPublishMethodBuilder
Publish a layer asset.

LayerUnpublishMethodBuilder
Unpublish a layer asset.

LayersListResponse
The response returned by a call to layers.List. Note: The list response does not include all the fields available in a layer. Refer to the layer resource description for details of the fields that are not included. You'll need to send a get request to retrieve the additional fields for each layer.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

LineStyle
Style for lines.

This type is not used in any activity, and only used as part of another schema.

LineStyleStroke
Stroke of the line.

This type is not used in any activity, and only used as part of another schema.

Map
A Map is a collection of Layers, optionally contained within folders.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

getPublished (none)

get (response)

create (request|response)

listPublished (none)

list (none)

publish (none)

patch (request)

permissions.list (none)

permissions.batchDelete (none)

unpublish (none)

permissions.batchUpdate (none)

delete (none)

MapContents
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

The contained type is Vec<MapItem>.

MapCreateMethodBuilder
Create a map asset.

MapDeleteMethodBuilder
Delete a map.

MapFolder
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

MapGetMethodBuilder
Return metadata for a particular map.

MapGetPublishedMethodBuilder
Return the published metadata for a particular map.

MapItem
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

MapKmlLink
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

MapLayer
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

MapListMethodBuilder
Return all maps readable by the current user.

MapListPublishedMethodBuilder
Return all published maps readable by the current user.

MapMethodsBuilder
A builder providing access to all methods supported on map resources. It is not used directly, but through the MapsEngine hub.

MapPatchMethodBuilder
Mutate a map asset.

MapPermissionBatchDeleteMethodBuilder
Remove permission entries from an already existing asset.

MapPermissionBatchUpdateMethodBuilder
Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

MapPermissionListMethodBuilder
Return all of the permissions for the specified asset.

MapPublishMethodBuilder
Publish a map asset.

MapUnpublishMethodBuilder
Unpublish a map asset.

MapsEngine
Central instance to access all MapsEngine related resource activities

MapsListResponse
The response returned by a call to maps.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

Parent
A list of the parents of an asset.

This type is not used in any activity, and only used as part of another schema.

ParentsListResponse
The response returned by a call to parents.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

parents.list (response)

parents.list (response)

parents.list (response)

parents.list (response)

parents.list (response)

Permission
A permission defines the user or group that has access to an asset, and the type of access they have.

This type is not used in any activity, and only used as part of another schema.

PermissionsBatchDeleteRequest
The request sent to mapsengine.permissions.batchDelete.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.batchDelete (request)

permissions.batchDelete (request)

permissions.batchDelete (request)

permissions.batchDelete (request)

permissions.batchDelete (request)

PermissionsBatchDeleteResponse
The response returned by a call to mapsengine.permissions.batchDelete.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.batchDelete (response)

permissions.batchDelete (response)

permissions.batchDelete (response)

permissions.batchDelete (response)

permissions.batchDelete (response)

PermissionsBatchUpdateRequest
The request sent to mapsengine.permissions.batchUpdate.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.batchUpdate (request)

permissions.batchUpdate (request)

permissions.batchUpdate (request)

permissions.batchUpdate (request)

permissions.batchUpdate (request)

PermissionsBatchUpdateResponse
The response returned by a call to mapsengine.permissions.batchUpdate.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.batchUpdate (response)

permissions.batchUpdate (response)

permissions.batchUpdate (response)

permissions.batchUpdate (response)

permissions.batchUpdate (response)

PermissionsListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.list (response)

permissions.list (response)

permissions.list (response)

permissions.list (response)

permissions.list (response)

permissions.list (response)

PointStyle
Style for points.

This type is not used in any activity, and only used as part of another schema.

PolygonStyle
Style for polygons.

This type is not used in any activity, and only used as part of another schema.

ProcessResponse
The response returned by a call to any asset's Process method.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

process (response)

process (response)

cancelProcessing (response)

process (response)

cancelProcessing (response)

process (response)

Project
A Maps Engine project groups a collection of resources.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (none)

icons.list (none)

icons.create (none)

icons.get (none)

ProjectIconCreateMethodBuilder
Create an icon.

ProjectIconGetMethodBuilder
Return an icon or its associated metadata

ProjectIconListMethodBuilder
Return all icons in the current project

ProjectListMethodBuilder
Return all projects readable by the current user.

ProjectMethodsBuilder
A builder providing access to all methods supported on project resources. It is not used directly, but through the MapsEngine hub.

ProjectsListResponse
The response returned by a call to projects.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

PublishResponse
The response returned by a call to any asset's Publish method.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

unpublish (response)

publish (response)

unpublish (response)

publish (response)

PublishedLayer
The published version of a layer.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

getPublished (response)

PublishedLayersListResponse
The response returned by a call to layers.List.published.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listPublished (response)

PublishedMap
The published version of a map asset.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

getPublished (response)

PublishedMapsListResponse
The response returned by a call to maps.List.published.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listPublished (response)

Raster
A geo-referenced raster.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.batchUpdate (none)

patch (request)

process (none)

permissions.list (none)

delete (none)

get (response)

list (none)

permissions.batchDelete (none)

parents.list (none)

upload (request|response)

files.insert (none)

RasterCollection
A raster collection groups multiple Raster resources for inclusion in a Layer.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

create (request|response)

parents.list (none)

permissions.list (none)

list (none)

process (none)

permissions.batchUpdate (none)

get (response)

permissions.batchDelete (none)

patch (request)

cancelProcessing (none)

rasters.batchInsert (none)

rasters.batchDelete (none)

delete (none)

rasters.list (none)

RasterCollectionCancelProcessingMethodBuilder
Cancel processing on a raster collection asset.

RasterCollectionCreateMethodBuilder
Create a raster collection asset.

RasterCollectionDeleteMethodBuilder
Delete a raster collection.

RasterCollectionGetMethodBuilder
Return metadata for a particular raster collection.

RasterCollectionListMethodBuilder
Return all raster collections readable by the current user.

RasterCollectionMethodsBuilder
A builder providing access to all methods supported on rasterCollection resources. It is not used directly, but through the MapsEngine hub.

RasterCollectionParentListMethodBuilder
Return all parent ids of the specified raster collection.

RasterCollectionPatchMethodBuilder
Mutate a raster collection asset.

RasterCollectionPermissionBatchDeleteMethodBuilder
Remove permission entries from an already existing asset.

RasterCollectionPermissionBatchUpdateMethodBuilder
Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

RasterCollectionPermissionListMethodBuilder
Return all of the permissions for the specified asset.

RasterCollectionProcesMethodBuilder
Process a raster collection asset.

RasterCollectionRasterBatchDeleteMethodBuilder
Remove rasters from an existing raster collection.

Up to 50 rasters can be included in a single batchDelete request. Each batchDelete request is atomic.

RasterCollectionRasterBatchInsertMethodBuilder
Add rasters to an existing raster collection. Rasters must be successfully processed in order to be added to a raster collection.

Up to 50 rasters can be included in a single batchInsert request. Each batchInsert request is atomic.

RasterCollectionRasterListMethodBuilder
Return all rasters within a raster collection.

RasterCollectionsListResponse
The response returned by a call to raster_collections.List. Note: The list response does not include all the fields available in a raster collection. Refer to the RasterCollection resource description for details of the fields that are not included. You'll need to send a get request to retrieve the additional fields for each raster collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

RasterCollectionsRaster
A raster resource.

This type is not used in any activity, and only used as part of another schema.

RasterCollectionsRasterBatchDeleteRequest
The request sent to rasterCollections.Rasters.BatchDelete.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

rasters.batchDelete (request)

RasterCollectionsRastersBatchDeleteResponse
The response returned by a call to rasterCollections.rasters.batchDelete.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

rasters.batchDelete (response)

RasterCollectionsRastersBatchInsertRequest
The request sent to rasterCollections.Rasters.BatchInsert.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

rasters.batchInsert (request)

RasterCollectionsRastersBatchInsertResponse
The response returned by a call to rasterCollections.rasters.batchInsert.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

rasters.batchInsert (response)

RasterCollectionsRastersListResponse
The response returned by a call to rasterCollections.rasters.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

rasters.list (response)

RasterDeleteMethodBuilder
Delete a raster.

RasterFileInsertMethodBuilder
Upload a file to a raster asset.

RasterGetMethodBuilder
Return metadata for a single raster.

RasterListMethodBuilder
Return all rasters readable by the current user.

RasterMethodsBuilder
A builder providing access to all methods supported on raster resources. It is not used directly, but through the MapsEngine hub.

RasterParentListMethodBuilder
Return all parent ids of the specified rasters.

RasterPatchMethodBuilder
Mutate a raster asset.

RasterPermissionBatchDeleteMethodBuilder
Remove permission entries from an already existing asset.

RasterPermissionBatchUpdateMethodBuilder
Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

RasterPermissionListMethodBuilder
Return all of the permissions for the specified asset.

RasterProcesMethodBuilder
Process a raster asset.

RasterUploadMethodBuilder
Create a skeleton raster asset for upload.

RastersListResponse
The response returned by a call to rasters.List.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

ScaledShape
Parameters for styling points as scaled shapes.

This type is not used in any activity, and only used as part of another schema.

ScalingFunction
Parameters for scaling scaled shapes.

This type is not used in any activity, and only used as part of another schema.

Schema
A schema indicating the properties which may be associated with features within a Table, and the types of those properties.

This type is not used in any activity, and only used as part of another schema.

SizeRange
Scaled shape size range in pixels. For circles, size corresponds to diameter.

This type is not used in any activity, and only used as part of another schema.

Table
A collection of geographic features, and associated metadata.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

features.batchInsert (none)

list (none)

patch (request)

get (response)

files.insert (none)

features.batchDelete (none)

permissions.batchDelete (none)

permissions.batchUpdate (none)

features.list (none)

process (none)

upload (request|response)

delete (none)

parents.list (none)

features.batchPatch (none)

permissions.list (none)

features.get (none)

create (request|response)

TableColumn
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

TableCreateMethodBuilder
Create a table asset.

TableDeleteMethodBuilder
Delete a table.

TableFeatureBatchDeleteMethodBuilder
Delete all features matching the given IDs.

TableFeatureBatchInsertMethodBuilder
Append features to an existing table.

A single batchInsert request can create:

Up to 50 features.

A combined total of 10000 vertices. Feature limits are documented in the Supported data formats and limits article of the Google Maps Engine help center. Note that free and paid accounts have different limits.

For more information about inserting features, read Creating features in the Google Maps Engine developer's guide.

TableFeatureBatchPatchMethodBuilder
Update the supplied features.

A single batchPatch request can update:

Up to 50 features.

A combined total of 10000 vertices. Feature limits are documented in the Supported data formats and limits article of the Google Maps Engine help center. Note that free and paid accounts have different limits.

Feature updates use HTTP PATCH semantics:

A supplied value replaces an existing value (if any) in that field.

Omitted fields remain unchanged.

Complex values in geometries and properties must be replaced as atomic units. For example, providing just the coordinates of a geometry is not allowed; the complete geometry, including type, must be supplied.

Setting a property's value to null deletes that property. For more information about updating features, read Updating features in the Google Maps Engine developer's guide.

TableFeatureGetMethodBuilder
Return a single feature, given its ID.

TableFeatureListMethodBuilder
Return all features readable by the current user.

TableFileInsertMethodBuilder
Upload a file to a placeholder table asset. See Table Upload in the Developer's Guide for more information. Supported file types are listed in the Supported data formats and limits article of the Google Maps Engine help center.

TableGetMethodBuilder
Return metadata for a particular table, including the schema.

TableListMethodBuilder
Return all tables readable by the current user.

TableMethodsBuilder
A builder providing access to all methods supported on table resources. It is not used directly, but through the MapsEngine hub.

TableParentListMethodBuilder
Return all parent ids of the specified table.

TablePatchMethodBuilder
Mutate a table asset.

TablePermissionBatchDeleteMethodBuilder
Remove permission entries from an already existing asset.

TablePermissionBatchUpdateMethodBuilder
Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

TablePermissionListMethodBuilder
Return all of the permissions for the specified asset.

TableProcesMethodBuilder
Process a table asset.

TableUploadMethodBuilder
Create a placeholder table asset to which table files can be uploaded. Once the placeholder has been created, files are uploaded to the https://www.googleapis.com/upload/mapsengine/v1/tables/table_id/files endpoint. See Table Upload in the Developer's Guide or Table.files: insert in the reference documentation for more information.

TablesListResponse
The response returned by a call to tables.List. Note: The list response does not include all the fields available in a table. Refer to the table resource description for details of the fields that are not included. You'll need to send a get request to retrieve the additional fields for each table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

Tags
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

The contained type is Vec<String>.

ValueRange
Range of values used for scaling shapes. The min/max values will be drawn as shapes with the min/max size.

This type is not used in any activity, and only used as part of another schema.

VectorStyle
A vector style contains styling information for vector layer.

This type is not used in any activity, and only used as part of another schema.

ZoomLevels
Zoom level range. Zoom levels are restricted between 0 and 24, inclusive.

This type is not used in any activity, and only used as part of another schema.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.